organization/requirements.txt
Django==3.2.6
gunicorn==20.0.4
psycopg2-binary==2.8.5
djangorestframework==3.12.4
djangorestframework_simplejwt==4.7.2
requests==2.26.0
unidecode==1.2.0